almost working

+++ src/server/Services/SearchLearning/SearchLearningService.cs
@@ -91,9 +91,9 @@ namespace Services.SearchLearning
             NeuralTestData data = new NeuralTestData();
 
             data.xs = items.Select(i => new List<double> {
-                i.Item.height / 1000,
-                i.Item.width / 1000,
-                i.Item.weight / 1000
+                (double)i.Item.height / 1000,
+                (double)i.Item.width / 1000,
+                (double)i.Item.weight / 1000
             });
 
             return data;
@@ -105,9 +105,9 @@ namespace Services.SearchLearning
             NeuralTrainingData data = new NeuralTrainingData();
 
             data.xs = itemsToUse.Select(i => new List<double> {
-                i.Item.height / 1000,
-                i.Item.width / 1000,
-                i.Item.weight / 1000
+                (double)i.Item.height / 1000,
+                (double)i.Item.width / 1000,
+                (double)i.Item.weight / 1000
             });
 
             data.ys = itemsToUse.Select(i => (double)i.LeadInQuery);
